# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Acoru Business Platform Baseline API
  version: 0.1.0
servers:
  - url: http://localhost:3000
paths:
...
  /api/v1/roles:
    get:
      summary: List roles
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Role list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Role"
                  meta:
                    $ref: "#/components/schemas/ListMeta"
    post:
      summary: Create a role
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateRoleRequest"
      responses:
        "201":
          description: Role created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Role"
...
  /api/v1/organizations:
    get:
      summary: List organizations
      parameters:
        - in: query
          name: q
          schema:
            type: string
        - in: query
          name: status
          schema:
            type: string
            enum: [active, inactive]
        - in: query
          name: limit
          schema:
            type: integer
            default: 50
        - in: query
          name: offset
          schema:
            type: integer
            default: 0
      responses:
        "200":
          description: Organization list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/Organization"
                  meta:
                    $ref: "#/components/schemas/ListMeta"
    post:
      summary: Create organization
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateOrganizationRequest"
      responses:
        "201":
          description: Organization created
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
  /api/v1/organizations/{id}:
    get:
      summary: Get organization
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      responses:
        "200":
          description: Organization detail
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
    patch:
      summary: Update organization
      parameters:
        - $ref: "#/components/parameters/ResourceId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOrganizationRequest"
      responses:
        "200":
          description: Organization updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: "#/components/schemas/Organization"
...
components:
...
  parameters:
    ResourceId:
      in: path
      name: id
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    ListMeta:
      type: object
      properties:
        limit:
          type: integer
        offset:
          type: integer
...
    Role:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        created_at: { type: string }
    CreateRoleRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        description: { type: string }
...
    Organization:
      type: object
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [active, inactive] }
        created_at: { type: string }
        updated_at: { type: string }
    CreateOrganizationRequest:
      type: object
      required: [code, name]
      properties:
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [active, inactive], default: active }
    UpdateOrganizationRequest:
      type: object
      properties:
        code: { type: string }
        name: { type: string }
        status: { type: string, enum: [active, inactive] }
...